Machine_Learning/HHOSSA-Tools/HHOSSA-Advanced Persistent Threat (APT) Detection System/requirements.txt
numpy
pandas
scikit-learn
lightgbm
tensorflow
flask
matplotlib
plotly
imblearn
kafka-python
